Adopt LWG-3546
I did split out transform_view::iterator tests

Addresses #1965
