+++ stl/inc/ranges
@@ -1127,7 +1127,7 @@ namespace ranges {
 #endif // _ITERATOR_DEBUG_LEVEL != 0
             }
 
-            _NODISCARD constexpr iterator_t<_Vw> base() const& noexcept(is_nothrow_copy_constructible_v<
+            _NODISCARD constexpr const iterator_t<_Vw>& base() const& noexcept(is_nothrow_copy_constructible_v<
                 iterator_t<_Vw>>) /* strengthened */ requires copyable<iterator_t<_Vw>> {
                 return _Current;
             }
@@ -1433,7 +1433,7 @@ namespace ranges {
                 : _Current{_STD move(_It._Current)}, _Parent{_It._Parent} {}
             // clang-format on
 
-            _NODISCARD constexpr iterator_t<_Base> base() const& noexcept(is_nothrow_copy_constructible_v<
+            _NODISCARD constexpr const iterator_t<_Base>& base() const& noexcept(is_nothrow_copy_constructible_v<
                 iterator_t<_Base>>) /* strengthened */ requires copyable<iterator_t<_Base>> {
                 return _Current;
             }
@@ -3775,7 +3775,7 @@ namespace ranges {
                 : _Current{_STD move(_It._Current)} {}
             // clang-format on
 
-            _NODISCARD constexpr iterator_t<_Base> base() const& noexcept(
+            _NODISCARD constexpr const iterator_t<_Base>& base() const& noexcept(
                 is_nothrow_copy_constructible_v<iterator_t<_Base>>) /* strengthened */
                 requires copyable<iterator_t<_Base>> {
                 return _Current;
