+++ tests/std/test.lst
@@ -385,6 +385,7 @@ tests\P0896R4_views_elements
 tests\P0896R4_views_empty
 tests\P0896R4_views_filter
 tests\P0896R4_views_filter_death
+tests\P0896R4_views_filter_iterator
 tests\P0896R4_views_iota
 tests\P0896R4_views_join
 tests\P0896R4_views_reverse
@@ -395,6 +396,7 @@ tests\P0896R4_views_take_while
 tests\P0896R4_views_take_while_death
 tests\P0896R4_views_transform
 tests\P0896R4_views_transform_death
+tests\P0896R4_views_transform_iterator
 tests\P0898R3_concepts
 tests\P0898R3_identity
 tests\P0912R5_coroutine
